requirement.txt
absl-py                      2.1.0
altair                       5.3.0
astunparse                   1.6.3
attrs                        23.2.0
blinker                      1.8.2
cachetools                   5.3.3
certifi                      2024.6.2
charset-normalizer           3.3.2
click                        8.1.7
flatbuffers                  24.3.25
gast                         0.6.0
gitdb                        4.0.11
GitPython                    3.1.43
google-pasta                 0.2.0
grpcio                       1.64.1
h5py                         3.11.0
idna                         3.7
Jinja2                       3.1.4
jsonschema                   4.22.0
jsonschema-specifications    2023.12.1
keras                        3.4.1
libclang                     18.1.1
Markdown                     3.6
markdown-it-py               3.0.0
MarkupSafe                   2.1.5
mdurl                        0.1.2
ml-dtypes                    0.3.2
namex                        0.0.8
numpy                        1.26.4
opencv-python-headless       4.10.0.84
opt-einsum                   3.3.0
optree                       0.11.0
packaging                    24.1
pandas                       2.2.2
pillow                       10.3.0
pip                          23.2.1
protobuf                     4.25.3
pyarrow                      16.1.0
pydeck                       0.9.1
Pygments                     2.18.0
python-dateutil              2.9.0.post0
pytz                         2024.1
referencing                  0.35.1
requests                     2.32.3
rich                         13.7.1
rpds-py                      0.18.1
scipy                        1.14.0
setuptools                   68.1.2
six                          1.16.0
smmap                        5.0.1
streamlit                    1.36.0
tenacity                     8.4.2
tensorboard                  2.16.2
tensorboard-data-server      0.7.2
tensorflow                   2.16.2
tensorflow-io-gcs-filesystem 0.37.0
termcolor                    2.4.0
toml                         0.10.2
toolz                        0.12.1
tornado                      6.4.1
typing_extensions            4.12.2
tzdata                       2024.1
urllib3                      2.2.2
watchdog                     4.0.1
Werkzeug                     3.0.3
wheel                        0.43.0
wrapt                        1.16.0
